swingfinder/requirements.txt
streamlit
pandas
numpy
plotly
requests
python-dotenv
scikit-learn
textblob
tiingo